[script.timers] 2.0.1 - Bugfix: actions in settings dialog don't work
